requirements.txt
# App / data
streamlit==1.54.0
pandas==3.0.1
numpy==2.3.1

# HF datasets / embeddings
datasets==4.6.1
sentence-transformers==5.2.3

# Retrieval / reranking
pyserini==0.44.0

# Model stack
torch==2.10.0
transformers==5.2.0

# RAG evaluation
ragas==0.4.3
litellm==1.82.0

# Tracing + orchestration
langfuse==3.3.3
langgraph==1.0.10
langchain-core==1.2.16

# AWS
boto3==1.35.99

# env
python-dotenv==1.0.1